style: edit inventoryView stylings and added navigation title
